Module jumbuck_nucleus::memory

Memory management implementation

SV39 page-based virtual-memory architecture for RV64 systems, and everything about memory management, like frame allocator, page table, map area and memory set, is implemented here.

Every task or process has a memory_set to control its virtual memory.

Modules

address ğŸ”’

Implementation of physical and virtual address and page number.

frame_allocator ğŸ”’

Implementation of FrameAllocator which controls all the frames in the operating system.

heap_allocator ğŸ”’

The global allocator

memory_set ğŸ”’

Implementation of MapArea and MemorySet.

page_table ğŸ”’

Implementation of PageTableEntry and PageTable.

Structs

FrameBlockTracker

manage a frame which has the same lifecycle as the tracker

KERNEL_SPACE

a memory set instance through lazy_static! managing kernel space

MapPermission

map permission corresponding to that in pte: R W X U

MemorySet

memory set structure, controls virtual-memory space

PageTableEntry

page table entry structure

PhysAddr

physical address ç‰©ç†åœ°å€çš„æŠ½è±¡è¡¨ç¤ºï¼Œä¸ºäº†å®‰å…¨çš„ç±»å‹æ£€æŸ¥

PhysPageNum

physical page number

VirtAddr

virtual address

VirtPageNum

virtual page number

Functions

alloc_one_frame

allocate a frame æ¥å£æš´éœ²ï¼šæœ¬æ¨¡å—å¤–çš„æ¥å£å…è®¸è°ƒç”¨allocæ–¹æ³•ï¼Œä½†æ˜¯ä¸å…è®¸è°ƒç”¨deallocã€‚å› ä¸ºä½¿ç”¨FrameBlockTrackeræ¥ç®¡ç†ç”Ÿå‘½å‘¨æœŸã€‚

init

initiate heap allocator, frame allocator and kernel space

remap_test

æ£€æŸ¥é¡µè¡¨æ˜¯å¦æ­£ç¡®è¿è¡Œã€‚

test_frame_allocator

ç‰©ç†é¡µå¸ åˆ†é…å™¨ æ€èƒ½æµ‹è¯•ã€‚

test_heap

æµ‹è¯•å †åˆ†é…å™¨æ˜¯å¦æ­£ç¡®å·¥ä½œ

test_panic_when_heap_space_not_enough

é€šè¿‡å †ä¸Šä½¿ç”¨è¿‡å¤šå†…å­˜ï¼Œè®©ç³»ç»Ÿå¥”æºƒ

translated_byte_buffer

translate a pointer to a mutable u8 Vec through page table

translated_refmut

translate a generic through page table and return a mutable reference

translated_str

translate a pointer to a mutable u8 Vec end with \0 through page table to a String