Module jumbuck_nucleus::memory::frame_allocator

Implementation of FrameAllocator which controls all the frames in the operating system.

Modules

frame_allocator_algorithm ğŸ”’

Structs

FRAME_ALLOCATOR

æŒ‰éœ€åˆå‹åŒ–çš„å…¨å±€ç‰©ç†é¡µå¸åˆ†é…å™¨ã€‚

FrameBlockTracker

manage a frame which has the same lifecycle as the tracker

Traits

FrameAllocator ğŸ”’

ç‰©ç†é¡µå¸çš„åˆ†é…å™¨æ¥å£ã€‚

Functions

alloc_frames

åˆ†é…ä¸€ç³»åˆ—è¿ç»­é¡µé¢ã€‚ æ¥å£æš´éœ²ï¼šæœ¬æ¨¡å—å¤–çš„æ¥å£å…è®¸è°ƒç”¨allocæ–¹æ³•ï¼Œä½†æ˜¯ä¸å…è®¸è°ƒç”¨deallocã€‚å› ä¸ºä½¿ç”¨FrameBlockTrackeræ¥ç®¡ç†ç”Ÿå‘½å‘¨æœŸã€‚

alloc_one_frame

allocate a frame æ¥å£æš´éœ²ï¼šæœ¬æ¨¡å—å¤–çš„æ¥å£å…è®¸è°ƒç”¨allocæ–¹æ³•ï¼Œä½†æ˜¯ä¸å…è®¸è°ƒç”¨deallocã€‚å› ä¸ºä½¿ç”¨FrameBlockTrackeræ¥ç®¡ç†ç”Ÿå‘½å‘¨æœŸã€‚

dealloc_frames ğŸ”’

get_remain_frame_cnt

è¿”å›å½“å‰å‰©ä¸‹çš„ç‰©ç†é¡µå¸æ•°é‡ã€‚

init_frame_allocator

initiate the frame allocator using end_kernel and MEMORY_END

test_frame_allocator

ç‰©ç†é¡µå¸ åˆ†é…å™¨ æ€èƒ½æµ‹è¯•ã€‚